requirements.txt
httpx>=0.27.0
anyio>=4.4.0
cachetools>=5.3.0
rich>=13.7.0
typer>=0.12.0
textblob>=0.18.0
feedparser>=6.0.11
python-dotenv>=1.0.0
numpy>=1.26.0
